rearrange bibdetails types and preprocessing
