+++ src/models/SearchResultsBib.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from "../types/searchTypes"
+import type { Bib } from "../types/bibTypes"
 import type { ElectronicResource } from "../types/bibTypes"
 import type { JSONLDValue } from "../types/itemTypes"
 import Item from "../models/Item"
@@ -26,7 +26,7 @@ export default class SearchResultsBib {
   numPhysicalItems: number
   items: Item[]
 
-  constructor(result: SearchResult) {
+  constructor(result: Bib) {
     this.id = result["@id"] ? result["@id"].substring(4) : ""
     this.title = this.getTitleFromResult(result)
     this.yearPublished = this.getYearFromResult(result)
@@ -89,7 +89,7 @@ export default class SearchResultsBib {
     }`
   }
 
-  getTitleFromResult(result: SearchResult) {
+  getTitleFromResult(result: Bib) {
     if (!result.titleDisplay || !result.titleDisplay.length) {
       const author =
         result.creatorLiteral && result.creatorLiteral.length
@@ -102,7 +102,7 @@ export default class SearchResultsBib {
     return result.titleDisplay[0]
   }
 
-  getYearFromResult(result: SearchResult) {
+  getYearFromResult(result: Bib) {
     const { dateStartYear, dateEndYear } = result
 
     const displayStartYear: string =
@@ -119,7 +119,7 @@ export default class SearchResultsBib {
   }
 
   // Map Bib items to Item class instances and sort them by their sortableShelfMark field
-  getItemsFromResult(result: SearchResult): Item[] {
+  getItemsFromResult(result: Bib): Item[] {
     return result.items
       .map((item) => {
         return new Item(item, this)
